+++ src/EPSG/Import/EPSGCodegenFromDataImport.php
@@ -8,6 +8,7 @@ declare(strict_types=1);
 
 namespace PHPCoord\EPSG\Import;
 
+use Exception;
 use PHPCoord\CoordinateOperation\CoordinateOperationMethods;
 use PHPCoord\CoordinateReferenceSystem\Compound;
 use PHPCoord\CoordinateReferenceSystem\CoordinateReferenceSystem;
@@ -28,6 +29,7 @@ use PHPCoord\UnitOfMeasure\Length\Length;
 use PHPCoord\UnitOfMeasure\Rate;
 use PHPCoord\UnitOfMeasure\Scale\Scale;
 use PHPCoord\UnitOfMeasure\Time\Time;
+use ReflectionClass;
 use SQLite3;
 
 use function array_map;
@@ -49,15 +51,29 @@ use function str_replace;
 use function ucwords;
 use function unlink;
 use function var_export;
+use function min;
+use function max;
+use function array_column;
+use function json_decode;
+use function substr;
+use function basename;
+use function class_exists;
+use function filemtime;
+use function touch;
+use function strtotime;
 
 use const SQLITE3_ASSOC;
 use const SQLITE3_OPEN_READONLY;
+use const JSON_THROW_ON_ERROR;
+use const PHP_EOL;
 
 class EPSGCodegenFromDataImport
 {
     private string $sourceDir;
 
-    private SQLite3 $sqlite;
+    private SQLite3 $dataDB;
+
+    private SQLite3 $extentDB;
 
     private Codegen $codeGen;
 
@@ -656,11 +672,17 @@ class EPSGCodegenFromDataImport
         $this->sourceDir = dirname(__DIR__, 2);
         $this->codeGen = new Codegen();
 
-        $this->sqlite = new SQLite3(
+        $this->dataDB = new SQLite3(
             __DIR__ . '/../../../resources/epsg/epsg.sqlite',
             SQLITE3_OPEN_READONLY
         );
-        $this->sqlite->enableExceptions(true);
+        $this->dataDB->enableExceptions(true);
+
+        $this->extentDB = new SQLite3(
+            __DIR__ . '/../../../resources/epsg/extents.sqlite',
+            SQLITE3_OPEN_READONLY
+        );
+        $this->extentDB->enableExceptions(true);
     }
 
     public function generateDataUnitsOfMeasure(): void
@@ -684,7 +706,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $data[$row['urn']] = $row;
@@ -721,7 +743,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $data[$row['urn']] = $row;
@@ -752,7 +774,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $data[$row['urn']] = $row;
@@ -782,7 +804,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $data[$row['urn']] = $row;
@@ -812,7 +834,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $data[$row['urn']] = $row;
@@ -841,7 +863,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $data[$row['urn']] = $row;
@@ -869,7 +891,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $data[$row['urn']] = $row;
@@ -900,7 +922,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
         ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             // some ellipsoids are defined via inverse flattening and the DB doesn't store the calculated data...
@@ -942,7 +964,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
         ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             if ($row['type'] === Datum::DATUM_TYPE_ENSEMBLE) {
@@ -957,7 +979,7 @@ class EPSGCodegenFromDataImport
                     ORDER BY d.datum_sequence
                     ";
 
-                $ensembleResult = $this->sqlite->query($ensembleSql);
+                $ensembleResult = $this->dataDB->query($ensembleSql);
                 while ($ensembleRow = $ensembleResult->fetchArray(SQLITE3_ASSOC)) {
                     $row['ensemble'][] = $ensembleRow['datum'];
                 }
@@ -1002,7 +1024,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $row['axes'] = [];
@@ -1019,7 +1041,7 @@ class EPSGCodegenFromDataImport
                 ORDER BY a.coord_axis_order
                 ";
 
-            $axisResult = $this->sqlite->query($axisSql);
+            $axisResult = $this->dataDB->query($axisSql);
             while ($axisRow = $axisResult->fetchArray(SQLITE3_ASSOC)) {
                 unset($axisRow['coord_sys']);
                 $row['axes'][] = $axisRow;
@@ -1051,7 +1073,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $row['axes'] = [];
@@ -1068,7 +1090,7 @@ class EPSGCodegenFromDataImport
                 ORDER BY a.coord_axis_order
                 ";
 
-            $axisResult = $this->sqlite->query($axisSql);
+            $axisResult = $this->dataDB->query($axisSql);
             while ($axisRow = $axisResult->fetchArray(SQLITE3_ASSOC)) {
                 unset($axisRow['coord_sys']);
                 $row['axes'][] = $axisRow;
@@ -1100,7 +1122,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $row['axes'] = [];
@@ -1117,7 +1139,7 @@ class EPSGCodegenFromDataImport
                 ORDER BY a.coord_axis_order
                 ";
 
-            $axisResult = $this->sqlite->query($axisSql);
+            $axisResult = $this->dataDB->query($axisSql);
             while ($axisRow = $axisResult->fetchArray(SQLITE3_ASSOC)) {
                 unset($axisRow['coord_sys']);
                 $row['axes'][] = $axisRow;
@@ -1149,7 +1171,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $row['axes'] = [];
@@ -1166,7 +1188,7 @@ class EPSGCodegenFromDataImport
                 ORDER BY a.coord_axis_order
                 ";
 
-            $axisResult = $this->sqlite->query($axisSql);
+            $axisResult = $this->dataDB->query($axisSql);
             while ($axisRow = $axisResult->fetchArray(SQLITE3_ASSOC)) {
                 unset($axisRow['coord_sys']);
                 $row['axes'][] = $axisRow;
@@ -1209,7 +1231,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $row['extent'] = array_values(array_unique(explode(',', $row['extent'])));
@@ -1258,7 +1280,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $row['extent'] = array_values(array_unique(explode(',', $row['extent'])));
@@ -1303,7 +1325,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $row['extent'] = array_values(array_unique(explode(',', $row['extent'])));
@@ -1349,7 +1371,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $row['extent'] = array_values(array_unique(explode(',', $row['extent'])));
@@ -1395,7 +1417,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $row['extent'] = array_values(array_unique(explode(',', $row['extent'])));
@@ -1444,7 +1466,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $row['extent'] = array_values(array_unique(explode(',', $row['extent'])));
@@ -1491,7 +1513,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $data[$row['urn']] = $row;
@@ -1533,7 +1555,7 @@ class EPSGCodegenFromDataImport
             ORDER BY name
         ';
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $row['reversible'] = (bool) $row['reversible'];
@@ -1551,7 +1573,7 @@ class EPSGCodegenFromDataImport
             ORDER BY u.sort_order ASC
         ";
 
-            $paramResult = $this->sqlite->query($sql);
+            $paramResult = $this->dataDB->query($sql);
             while ($paramRow = $paramResult->fetchArray(SQLITE3_ASSOC)) {
                 $paramName = self::makeParamName($paramRow['parameter_name']);
                 $data[$row['urn']]['paramData'][$paramName] = [
@@ -1625,7 +1647,7 @@ class EPSGCodegenFromDataImport
             ORDER BY source_crs, target_crs, operation
             ';
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $byRegion = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $extents = explode(',', $row['extent']);
@@ -1695,7 +1717,7 @@ class EPSGCodegenFromDataImport
             ORDER BY urn
             ';
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $row['extent'] = array_values(array_unique(explode(',', $row['extent'])));
@@ -1725,7 +1747,7 @@ class EPSGCodegenFromDataImport
                     ORDER BY pu.sort_order
                     ";
 
-            $paramsResult = $this->sqlite->query($paramsSql);
+            $paramsResult = $this->dataDB->query($paramsSql);
             while ($paramsRow = $paramsResult->fetchArray(SQLITE3_ASSOC)) {
                 unset($paramsRow['operation_code']);
                 $paramsRow['name'] = self::makeParamName($paramsRow['name']);
@@ -1787,6 +1809,218 @@ class EPSGCodegenFromDataImport
         $this->codeGen->generateSupportedOperations();
     }
 
+    public function generateExtents(): void
+    {
+        echo 'Updating extents...';
+
+        $boundingBoxOnly = $this->sourceDir . '/Geometry/Extents/BoundingBoxOnly/';
+        $builtInFull = $this->sourceDir . '/Geometry/Extents/';
+        $africa = $this->sourceDir . '/../vendor/php-coord/datapack-africa/src/Geometry/Extents/';
+        $antarctic = $this->sourceDir . '/../vendor/php-coord/datapack-antarctic/src/Geometry/Extents/';
+        $arctic = $this->sourceDir . '/../vendor/php-coord/datapack-arctic/src/Geometry/Extents/';
+        $asia = $this->sourceDir . '/../vendor/php-coord/datapack-asia/src/Geometry/Extents/';
+        $europe = $this->sourceDir . '/../vendor/php-coord/datapack-europe/src/Geometry/Extents/';
+        $northAmerica = $this->sourceDir . '/../vendor/php-coord/datapack-northamerica/src/Geometry/Extents/';
+        $southAmerica = $this->sourceDir . '/../vendor/php-coord/datapack-southamerica/src/Geometry/Extents/';
+        $oceania = $this->sourceDir . '/../vendor/php-coord/datapack-oceania/src/Geometry/Extents/';
+
+        $regionMap = (new RegionMap())();
+
+        $sql = "
+            SELECT e.extent_code, e.extent_name, e.revision_date
+            FROM epsg_coordinatereferencesystem crs
+            JOIN epsg_usage u ON u.object_code = crs.coord_ref_sys_code AND u.object_table_name = 'epsg_coordinatereferencesystem'
+            JOIN epsg_extent e ON u.extent_code = e.extent_code
+            LEFT JOIN epsg_deprecation dep ON dep.object_table_name = 'epsg_coordinatereferencesystem' AND dep.object_code = crs.coord_ref_sys_code AND dep.deprecation_date <= '2020-12-14'
+            WHERE dep.deprecation_id IS NULL AND e.deprecated = 0
+            AND crs.coord_ref_sys_kind NOT IN ('engineering', 'derived') AND crs.coord_ref_sys_name NOT LIKE '%example%' AND crs.coord_ref_sys_name NOT LIKE '%mining%'
+
+            UNION
+
+            SELECT e.extent_code, e.extent_name, e.revision_date
+            FROM epsg_coordoperation o
+            JOIN epsg_coordinatereferencesystem sourcecrs ON sourcecrs.coord_ref_sys_code = o.source_crs_code AND sourcecrs.coord_ref_sys_kind NOT IN ('engineering', 'derived') AND sourcecrs.deprecated = 0
+            JOIN epsg_usage u ON u.object_code = o.coord_op_code AND u.object_table_name = 'epsg_coordoperation'
+            JOIN epsg_extent e ON u.extent_code = e.extent_code
+            LEFT JOIN epsg_deprecation dep ON dep.object_table_name = 'epsg_coordoperation' AND dep.object_code = o.coord_op_code AND dep.deprecation_date <= '2020-12-14'
+            LEFT JOIN epsg_supersession s ON s.object_table_name = 'epsg_coordoperation' AND s.object_code = o.coord_op_code
+            WHERE dep.deprecation_id IS NULL AND e.deprecated = 0 AND s.supersession_id IS NULL
+            AND o.coord_op_type != 'conversion' AND o.coord_op_type != 'concatenated operation' AND o.coord_op_name NOT LIKE '%example%' AND o.coord_op_name NOT LIKE '%mining%'
+            AND o.coord_op_method_code NOT IN (" . implode(',', self::BLACKLISTED_METHODS) . ')
+            AND o.coord_op_code NOT IN (' . implode(',', $this->getBlacklistedOperations()) . ")
+
+            UNION
+
+            SELECT e.extent_code, e.extent_name, e.revision_date
+            FROM epsg_coordoperation o
+            JOIN epsg_coordinatereferencesystem projcrs ON projcrs.projection_conv_code = o.coord_op_code AND projcrs.coord_ref_sys_kind NOT IN ('engineering', 'derived') AND projcrs.deprecated = 0
+            JOIN epsg_usage u ON u.object_code = o.coord_op_code AND u.object_table_name = 'epsg_coordoperation'
+            JOIN epsg_extent e ON u.extent_code = e.extent_code
+            LEFT JOIN epsg_deprecation dep ON dep.object_table_name = 'epsg_coordoperation' AND dep.object_code = o.coord_op_code AND dep.deprecation_date <= '2020-12-14'
+            LEFT JOIN epsg_supersession s ON s.object_table_name = 'epsg_coordoperation' AND s.object_code = o.coord_op_code
+            WHERE dep.deprecation_id IS NULL AND e.deprecated = 0 AND s.supersession_id IS NULL
+            AND o.coord_op_type = 'conversion' AND o.coord_op_type != 'concatenated operation' AND o.coord_op_name NOT LIKE '%example%' AND o.coord_op_name NOT LIKE '%mining%'
+            AND o.coord_op_method_code NOT IN (" . implode(',', self::BLACKLISTED_METHODS) . ')
+            AND o.coord_op_code NOT IN (' . implode(',', $this->getBlacklistedOperations()) . ')
+
+            GROUP BY e.extent_code
+        ';
+        $result = $this->dataDB->query($sql);
+
+        $extents = [];
+        while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
+            $extents[$row['extent_code']] = $row;
+
+            if (!isset($regionMap[$row['extent_code']])) {
+                throw new Exception("Unknown region for {$row['extent_code']}:{$row['extent_name']}");
+            }
+        }
+
+        $this->codeGen->updateFileData($this->sourceDir . '/Geometry/Extents/ExtentMap.php', array_map(fn (array $extent) => ['name' => $extent['extent_name']], $extents));
+
+        foreach ($extents as $extentCode => $extent) {
+            if (class_exists("PHPCoord\\Geometry\\Extents\\Extent{$extentCode}")) {
+                $reflectionClass = new ReflectionClass("PHPCoord\\Geometry\\Extents\\Extent{$extentCode}");
+                if (filemtime($reflectionClass->getFileName()) === strtotime($extent['revision_date'])) {
+                    continue;
+                }
+            }
+            echo $extentCode . ' ' . $extent['extent_name'] . PHP_EOL;
+
+            $region = $regionMap[$extentCode];
+            $originalPolygons = json_decode($this->extentDB->querySingle("SELECT original FROM extent WHERE extent_code = {$extentCode}"), true, 512, JSON_THROW_ON_ERROR);
+            $exportSimple = "<?php\ndeclare(strict_types=1);\n\nnamespace PHPCoord\Geometry\Extents\BoundingBoxOnly;\n/**\n * {$region}/{$extent['extent_name']}.\n * @internal\n */\nclass Extent{$extentCode}\n{\n    public function __invoke(): array\n    {\n        return\n        [\n";
+
+            if ($originalPolygons['type'] === 'Polygon') {
+                $originalPolygons['coordinates'] = [$originalPolygons['coordinates']];
+            }
+
+            if (in_array($extentCode, [1262, 2346, 2830, 4520, 4523], true)) {
+                $originalPolygons['coordinates'] = [[[[-180, -90], [-180, 90], [180, 90], [180, -90], [-180, -90]]]]; // don't overcomplicate it!
+            }
+
+            foreach ($originalPolygons['coordinates'] as $polygon) {
+                $outerRingPoints = $polygon[0];
+                $xmin = min(array_column($outerRingPoints, 0));
+                $xmax = max(array_column($outerRingPoints, 0));
+                $ymin = min(array_column($outerRingPoints, 1));
+                $ymax = max(array_column($outerRingPoints, 1));
+                $exportSimple .= "            [\n                [\n                    ";
+                $exportSimple .= "[{$xmax}, {$ymax}], [{$xmin}, {$ymax}], [{$xmin}, {$ymin}], [{$xmax}, {$ymin}], [{$xmax}, {$ymax}],";
+                $exportSimple .= "\n                ],\n            ],\n";
+            }
+            $exportSimple .= "        ];\n    }\n}\n";
+
+            $bufferedPolygons = json_decode($this->extentDB->querySingle("SELECT buffered FROM extent WHERE extent_code = {$extentCode}"), true, 512, JSON_THROW_ON_ERROR);
+            $exportFull = "<?php\ndeclare(strict_types=1);\n\nnamespace PHPCoord\Geometry\Extents;\n/**\n * {$region}/{$extent['extent_name']}.\n * @internal\n */\nclass Extent{$extentCode}\n{\n    public function __invoke(): array\n    {\n        return\n        [\n";
+            if ($bufferedPolygons['type'] === 'Polygon') {
+                $bufferedPolygons['coordinates'] = [$bufferedPolygons['coordinates']];
+            }
+            if (in_array($extentCode, [1262, 2346, 2830, 4520, 4523], true)) {
+                $bufferedPolygons['coordinates'] = [[[[-180, -90], [-180, 90], [180, 90], [180, -90], [-180, -90]]]]; // don't overcomplicate it!
+            }
+            foreach ($bufferedPolygons['coordinates'] as $polygon) {
+                $exportFull .= "            [\n";
+
+                foreach ($polygon as $ring) {
+                    $exportFull .= "                [\n                    ";
+                    foreach ($ring as $point) {
+                        $exportFull .= '[' . $point[0] . ', ' . $point[1] . '], ';
+                    }
+                    $exportFull .= "\n                ],\n";
+                }
+                $exportFull .= "            ],\n";
+            }
+            $exportFull .= "        ];\n    }\n}\n";
+
+            switch ($region) {
+                case RegionMap::REGION_GLOBAL:
+                    file_put_contents($builtInFull . "Extent{$extentCode}.php", $exportFull);
+                    $this->codeGen->csFixFile($builtInFull . "Extent{$extentCode}.php");
+                    touch($builtInFull . "Extent{$extentCode}.php", strtotime($extent['revision_date']));
+                    break;
+                case RegionMap::REGION_AFRICA:
+                    file_put_contents($boundingBoxOnly . "Extent{$extentCode}.php", $exportSimple);
+                    $this->codeGen->csFixFile($boundingBoxOnly . "Extent{$extentCode}.php");
+                    touch($boundingBoxOnly . "Extent{$extentCode}.php", strtotime($extent['revision_date']));
+                    file_put_contents($africa . "Extent{$extentCode}.php", $exportFull);
+                    $this->codeGen->csFixFile($africa . "Extent{$extentCode}.php");
+                    touch($africa . "Extent{$extentCode}.php", strtotime($extent['revision_date']));
+                    break;
+                case RegionMap::REGION_ANTARCTIC:
+                    file_put_contents($boundingBoxOnly . "Extent{$extentCode}.php", $exportSimple);
+                    $this->codeGen->csFixFile($boundingBoxOnly . "Extent{$extentCode}.php");
+                    touch($boundingBoxOnly . "Extent{$extentCode}.php", strtotime($extent['revision_date']));
+                    file_put_contents($antarctic . "Extent{$extentCode}.php", $exportFull);
+                    $this->codeGen->csFixFile($antarctic . "Extent{$extentCode}.php");
+                    touch($antarctic . "Extent{$extentCode}.php", strtotime($extent['revision_date']));
+                    break;
+                case RegionMap::REGION_ARCTIC:
+                    file_put_contents($boundingBoxOnly . "Extent{$extentCode}.php", $exportSimple);
+                    $this->codeGen->csFixFile($boundingBoxOnly . "Extent{$extentCode}.php");
+                    touch($boundingBoxOnly . "Extent{$extentCode}.php", strtotime($extent['revision_date']));
+                    file_put_contents($arctic . "Extent{$extentCode}.php", $exportFull);
+                    $this->codeGen->csFixFile($arctic . "Extent{$extentCode}.php");
+                    touch($arctic . "Extent{$extentCode}.php", strtotime($extent['revision_date']));
+                    break;
+                case RegionMap::REGION_ASIA:
+                    file_put_contents($boundingBoxOnly . "Extent{$extentCode}.php", $exportSimple);
+                    $this->codeGen->csFixFile($boundingBoxOnly . "Extent{$extentCode}.php");
+                    touch($boundingBoxOnly . "Extent{$extentCode}.php", strtotime($extent['revision_date']));
+                    file_put_contents($asia . "Extent{$extentCode}.php", $exportFull);
+                    $this->codeGen->csFixFile($asia . "Extent{$extentCode}.php");
+                    touch($asia . "Extent{$extentCode}.php", strtotime($extent['revision_date']));
+                    break;
+                case RegionMap::REGION_OCEANIA:
+                    file_put_contents($boundingBoxOnly . "Extent{$extentCode}.php", $exportSimple);
+                    $this->codeGen->csFixFile($boundingBoxOnly . "Extent{$extentCode}.php");
+                    touch($boundingBoxOnly . "Extent{$extentCode}.php", strtotime($extent['revision_date']));
+                    file_put_contents($oceania . "Extent{$extentCode}.php", $exportFull);
+                    $this->codeGen->csFixFile($oceania . "Extent{$extentCode}.php");
+                    touch($oceania . "Extent{$extentCode}.php", strtotime($extent['revision_date']));
+                    break;
+                case RegionMap::REGION_EUROPE:
+                    file_put_contents($boundingBoxOnly . "Extent{$extentCode}.php", $exportSimple);
+                    $this->codeGen->csFixFile($boundingBoxOnly . "Extent{$extentCode}.php");
+                    touch($boundingBoxOnly . "Extent{$extentCode}.php", strtotime($extent['revision_date']));
+                    file_put_contents($europe . "Extent{$extentCode}.php", $exportFull);
+                    $this->codeGen->csFixFile($europe . "Extent{$extentCode}.php");
+                    touch($europe . "Extent{$extentCode}.php", strtotime($extent['revision_date']));
+                    break;
+                case RegionMap::REGION_NORTHAMERICA:
+                    file_put_contents($boundingBoxOnly . "Extent{$extentCode}.php", $exportSimple);
+                    $this->codeGen->csFixFile($boundingBoxOnly . "Extent{$extentCode}.php");
+                    touch($boundingBoxOnly . "Extent{$extentCode}.php", strtotime($extent['revision_date']));
+                    file_put_contents($northAmerica . "Extent{$extentCode}.php", $exportFull);
+                    $this->codeGen->csFixFile($northAmerica . "Extent{$extentCode}.php");
+                    touch($northAmerica . "Extent{$extentCode}.php", strtotime($extent['revision_date']));
+                    break;
+                case RegionMap::REGION_SOUTHAMERICA:
+                    file_put_contents($boundingBoxOnly . "Extent{$extentCode}.php", $exportSimple);
+                    $this->codeGen->csFixFile($boundingBoxOnly . "Extent{$extentCode}.php");
+                    touch($boundingBoxOnly . "Extent{$extentCode}.php", strtotime($extent['revision_date']));
+                    file_put_contents($southAmerica . "Extent{$extentCode}.php", $exportFull);
+                    $this->codeGen->csFixFile($southAmerica . "Extent{$extentCode}.php");
+                    touch($southAmerica . "Extent{$extentCode}.php", strtotime($extent['revision_date']));
+                    break;
+                default:
+                    throw new Exception("Unknown region: {$region}");
+            }
+        }
+
+        // Remove unused extents
+        foreach ([$boundingBoxOnly, $builtInFull, $africa, $antarctic, $arctic, $asia, $europe, $northAmerica, $southAmerica, $oceania] as $extentType) {
+            foreach (glob($extentType . '/Extent[0-9]*.php') as $filename) {
+                $code = substr(basename($filename, '.php'), 6);
+                if (!isset($extents[$code])) {
+                    unlink($filename);
+                }
+            }
+        }
+
+        echo 'done' . PHP_EOL;
+    }
+
     protected static function makeParamName(string $string): string
     {
         $string = str_replace([' ', '-', '(', ')', '"'], '', ucwords($string, ' -()"'));
@@ -1831,7 +2065,7 @@ class EPSGCodegenFromDataImport
             WHERE wgs84.target_crs_code IN (4978,4326,4979)
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $data[] = $row['coord_op_code'];
@@ -1850,7 +2084,7 @@ class EPSGCodegenFromDataImport
             WHERE etrs89.target_crs_code IN (4936, 4258, 4937)
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $data[] = $row['coord_op_code'];
@@ -1869,7 +2103,7 @@ class EPSGCodegenFromDataImport
             WHERE wgs84.target_crs_code IN (4978,4326,4979)
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $data[] = $row['coord_op_code'];
@@ -1888,7 +2122,7 @@ class EPSGCodegenFromDataImport
             WHERE wgs84.target_crs_code IN (4978,4326,4979)
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $data[] = $row['coord_op_code'];
@@ -1907,7 +2141,7 @@ class EPSGCodegenFromDataImport
             WHERE wgs84.target_crs_code IN (4978,4326,4979)
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $data[] = $row['coord_op_code'];
@@ -1925,7 +2159,7 @@ class EPSGCodegenFromDataImport
             JOIN epsg_coordinatereferencesystem t ON o.target_crs_code = t.coord_ref_sys_code AND t.coord_ref_sys_kind = 'projected'
             ";
 
-        $result = $this->sqlite->query($sql);
+        $result = $this->dataDB->query($sql);
         $data = [];
         while ($row = $result->fetchArray(SQLITE3_ASSOC)) {
             $data[] = $row['coord_op_code'];